+++ lizard_graph/views.py
@@ -84,13 +84,14 @@ def cached_time_series_from_graph_item(graph_item, start, end):
     Cached version of graph_item.time_series(start, end)
     """
     def time_series_key(graph_item, start, end):
-        return ('ts::%s:%s:%s:%s:%s:%s::%s:%s' % (
+        return ('ts::%s:%s:%s:%s:%s:%s:%s::%s:%s' % (
             graph_item.fews_norm_source.database_name, graph_item.location,
+            graph_item.related_location,
             graph_item.parameter, graph_item.module, graph_item.time_step,
             graph_item.qualifierset, start, end)).replace(' ', '_')
     cache_key = time_series_key(graph_item, start, end)
     ts = cache.get(cache_key)
-    if ts is None:
+    if 1:  # if ts is None:
         ts = graph_item.time_series(start, end)
         cache.set(cache_key, ts)
     return ts